+++ tablet_qt/core/camcopsapp.cpp
@@ -501,6 +501,8 @@ void CamcopsApp::patientRegistrationFailed(
     const NetworkManager::ErrorCode error_code, const QString& error_string
 )
 {
+    tidyUpFailedRegistration();
+
     deleteNetworkGuiGuard();
 
     const QString base_message
@@ -536,6 +538,18 @@ void CamcopsApp::patientRegistrationFailed(
     maybeRetryNetworkOperation(
         base_message, additional_message, NetworkOperation::RegisterPatient
     );
+}
+
+void CamcopsApp::tidyUpFailedRegistration()
+{
+    // An erratic firewall that rejects some genuine requests and allows others
+    // can result in broken registrations unless we tidy up properly. If
+    // registering the patient succeeds but registering the device fails,
+    // unless we delete the partly registered patient, we will end up with
+    // multiple patients in the database with the same ID and we won't be able
+    // to submit tasks. We're in single user mode so simplest just to delete
+    // all patients.
+    m_datadb->deleteFrom(Patient::TABLENAME);
 }
 
 void CamcopsApp::patientRegistrationFinished()